Display form on footnote link click
